requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
asyncpg==0.30.0
openpyxl==3.1.5
reportlab==4.4.1
alembic==1.13.1
greenlet==3.0.3

# Rate limiting
slowapi==0.1.9

# Monitoring
sentry-sdk[fastapi]==1.40.6

# AI Integration
openai==1.12.0
httpx==0.27.0
